Try to add tests
