+++ package.json
@@ -26,18 +26,25 @@
     "gatsby-plugin-react-helmet": "^3.2.3",
     "gatsby-plugin-react-svg": "^3.0.0",
     "gatsby-plugin-sharp": "^2.5.5",
+    "gsap": "^3.2.6",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-helmet": "^6.0.0",
+    "save": "^2.4.0",
     "tailwindcss": "^1.2.0",
     "uuid": "^7.0.3"
   },
   "devDependencies": {
     "babel-eslint": "^10.1.0",
+    "babel-jest": "^25.4.0",
+    "babel-preset-gatsby": "^0.3.3",
     "eslint": "^6.8.0",
     "eslint-config-prettier": "^6.10.1",
     "eslint-plugin-react": "^7.19.0",
-    "prettier": "^2.0.4"
+    "identity-obj-proxy": "^3.0.0",
+    "jest": "^25.4.0",
+    "prettier": "^2.0.4",
+    "react-test-renderer": "^16.13.1"
   },
   "repository": {
     "type": "git",
