+++ src/pages/index.js
@@ -2,15 +2,8 @@
 
 Inspirational resources: 
 
-https://www.npmjs.com/package/react-scrollmagic
-
-https://www.tailwindtoolbox.com/
-
-https://dev.to/sm0ke/tailwind-css-open-source-components-364h
-
-https://github.com/creativetimofficial/tailwind-starter-kit
-
-https://github.com/framer/motion
+https://www.dcode.no/
+https://www.coretrek.no/
 
 */
 
@@ -19,9 +12,9 @@ import React from "react"
 import Navbar from "../components/Header/Navbar.component"
 import NavbarContent from "../components/Header/NavbarContent.component"
 
-import MainContent from "../components/Index/MainContent.component"
+import MainContent from "../components/Design3/MainContent.component"
 
-function IndexPage() {
+function NyIndexPage() {
   return (
     <>
       <Navbar>
@@ -32,4 +25,4 @@ function IndexPage() {
   )
 }
 
-export default IndexPage
+export default NyIndexPage
